Don't assume BaseAddressRecord.ctnr exists
See issue #743 for details. In summary, Django catches and stores
ValidationErrors and raises them all at once, so BaseAddressRecord.clean
might be working with an invalid object; if it raises an exception other
than ValidationError, the exception will propagate up to the view
dispatcher, which will make the request return an HTTP 500 status code.
In this case, accessing the ctnr field might raise Ctnr.DoesNotExist.