Don't make maintain_sb query at module import
The -q/--qlobber option of maintain_migrate requires that dns_migrate be
imported. Because get_delegated is called when dns_migrate is imported,
it's possible that the maintain sandbox database doesn't exist (which
raises an exception) or that the data in maintain_sb is out-of-date
(which causes results inconsistent with the intent of -q/--qlobber).